updates to LockServiceImpl

+++ sample-app/reservation/src/main/java/com/ferenc/reservation/repository/lock/LockServiceImpl.java
@@ -4,10 +4,14 @@ import com.ferenc.reservation.repository.PessimisticLockRepository;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.dao.DataAccessException;
 import org.springframework.dao.DuplicateKeyException;
+import org.springframework.data.mongodb.MongoTransactionManager;
 import org.springframework.retry.annotation.Backoff;
 import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.TransactionStatus;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.Instant;
 import java.util.Date;
@@ -24,15 +28,29 @@ public class LockServiceImpl implements LockService {
 
     private final PessimisticLockRepository pessimisticLockRepository;
 
+    private final MongoTransactionManager transactionManager;
+
     @Override
-    @Retryable(value = {DuplicateKeyException.class}, maxAttempts = 4, backoff = @Backoff(delay = 1000))
+    @Transactional
+    @Retryable(value = {DataAccessException.class}, maxAttempts = 4, backoff = @Backoff(delay = 300))
     public boolean acquireLock(String licencePlate, String userId) {
-        PessimisticLock pessimisticLock = new PessimisticLock();
-        pessimisticLock.setId(licencePlate);
-        pessimisticLock.setUserId(userId);
-        pessimisticLockRepository.insert(pessimisticLock);
-        logger.info("Lock created: {}, at {}", licencePlate, pessimisticLock.getCreatedDate());
-        return true;
+        TransactionStatus transactionStatus = transactionManager.getTransaction(null);
+        PessimisticLock existingLock = pessimisticLockRepository.findById(licencePlate).orElse(null);
+        try{
+            if (existingLock != null) {
+                throw new DuplicateKeyException(String.format("Car %s currently not available.", licencePlate ));
+            }
+            PessimisticLock pessimisticLock = new PessimisticLock();
+            pessimisticLock.setId(licencePlate);
+            pessimisticLock.setUserId(userId);
+            pessimisticLockRepository.insert(pessimisticLock);
+            logger.info("Lock created: {}, at {}", licencePlate, pessimisticLock.getCreatedDate());
+            transactionManager.commit(transactionStatus);
+            return true;
+        } catch (Throwable e) {
+            transactionManager.rollback(transactionStatus);
+            throw e;
+        }
     }
 
     @Override
@@ -54,7 +72,9 @@ public class LockServiceImpl implements LockService {
         Instant expiration = now.minusMillis(EXPIRY_TIME_MS);
 
         List<PessimisticLock> locks = pessimisticLockRepository.findByCreatedDateBefore(Date.from(expiration));
-        pessimisticLockRepository.deleteAll(locks);
-        logger.info("Locks deleted: {}, at {}.", locks.size(), now);
+        if(locks.size()>0) {
+            pessimisticLockRepository.deleteAll(locks);
+            logger.info("Locks deleted: {}, at {}.", locks.size(), now);
+        }
     }
 }
